+++ lib/rspec/expectations/syntax.rb
@@ -69,10 +69,8 @@ module RSpec
         return if expect_enabled?(syntax_host)
 
         syntax_host.module_exec do
-          def expect(*target, &target_block)
-            target << target_block if block_given?
-            raise ArgumentError.new("You must pass an argument or a block to #expect but not both.") unless target.size == 1
-            ::RSpec::Expectations::ExpectationTarget.new(target.first)
+          def expect(value=::RSpec::Expectations::ExpectationTarget::UndefinedValue, &block)
+            ::RSpec::Expectations::ExpectationTarget.for(value, block)
           end
         end
       end